Add function to retrieve all pages of a query
This commit adds a new function `get_all_results` that retrieves all pages of a query from the API response. It uses the `meta` information of the response to calculate the total number of pages and iterates over each page, appending the results to a list. A progress bar is also added to track the retrieval progress.

Details:
- Install `tqdm` library
- Update imports
- Modify `get_all_results` function to accept an API function and additional parameters
- Update code in cell 3 to use `get_all_results` with appropriate arguments